\name{GetDetails-methods}
\docType{methods}
\alias{GetDetails-methods}
\alias{GetDetails}
\alias{AceEstimate-method}
\title{ A generic function for extracting the \code{Details} slot of an  object.}
\description{
A generic function for extracting the \code{Details} slot of an \code{AceEstimation} object.
}
\section{Methods}{
\describe{
Here's a filler description.
\item{\code{signature(object = "AceEstimate")}}{Extracts the \code{Details} slot of an \code{AceEstimation} object.}
}}
\author{
Will Beasley
}
\keyword{methods}
