\name{GetEstimate-methods}
\docType{methods}
\alias{GetEstimate-methods}
\alias{GetEstimate,AceEstimate-method}
\title{ Generic function for returning the contents from an \code{AceEstimate} class.}
\description{
 Extract the values
}
\section{Methods}{
\describe{

\item{\code{signature(object = "AceEstimate")}}{
%%  ~~describe this method here~~
}
}}
\author{
Will Beasley
}
\keyword{methods}

