\name{static-package}
\alias{static-package}
\docType{package}
\title{
  Static Local Variables
}
\description{
  Static local variables are initialized once,
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  Maintainer: Neal Fultz <nfultz@gmail.com>
}
\keyword{ package }
\seealso{
  \link{static}
}
